requirements.txt
# ================================================
# IntelliQuery AI - Requirements
# ================================================
# Production-Ready Intelligent Query Processing
# Install: pip install -r requirements.txt

# ---------- WEB FRAMEWORK ----------
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2

# ---------- DATABRICKS ----------
databricks-sql-connector>=3.0.0
databricks-vectorsearch>=0.22      # Vector search for semantic similarity
requests>=2.31.0

# ---------- DATA PROCESSING ----------
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0        # Excel file support
python-dotenv>=1.0.0   # .env file loading

# ---------- MACHINE LEARNING ----------
scikit-learn>=1.3.0
joblib>=1.3.0          # Model persistence

# ---------- VISUALIZATION ----------
matplotlib>=3.7.0

# ---------- DOCUMENT PROCESSING ----------
langchain>=0.1.0
langchain-community>=0.0.20
pypdf>=3.17.0
unstructured>=0.10.0
